# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Chally API",
    "description": "Chally API for peer accountability challenges",
    "version": "0.1.0"
  },
  "paths": {
...
    "/wallet": {
      "get": {
        "tags": [
          "wallet"
        ],
        "summary": "Get Wallet",
        "operationId": "get_wallet_wallet_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WalletSnapshot"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/wallet/deposit/checkout": {
      "post": {
        "tags": [
          "wallet"
        ],
        "summary": "Create Deposit Checkout",
        "operationId": "create_deposit_checkout_wallet_deposit_checkout_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDepositRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateDepositResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/wallet/withdraw/refund": {
      "post": {
        "tags": [
          "wallet"
        ],
        "summary": "Withdraw Refund",
        "description": "Withdraw tokens back to card via Stripe refunds (FIFO by deposit lots).",
        "operationId": "withdraw_refund_wallet_withdraw_refund_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WithdrawRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WithdrawResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateDepositRequest": {
        "properties": {
          "tokens": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Tokens",
            "description": "Number of tokens to buy"
          },
          "success_url": {
            "type": "string",
            "minLength": 1,
            "format": "uri",
            "title": "Success Url"
          },
          "cancel_url": {
            "type": "string",
            "minLength": 1,
            "format": "uri",
            "title": "Cancel Url"
          }
        },
        "type": "object",
        "required": [
          "tokens",
          "success_url",
          "cancel_url"
        ],
        "title": "CreateDepositRequest"
      },
      "CreateDepositResponse": {
        "properties": {
          "checkout_url": {
            "type": "string",
            "title": "Checkout Url"
          },
          "session_id": {
            "type": "string",
            "title": "Session Id"
          }
        },
        "type": "object",
        "required": [
          "checkout_url",
          "session_id"
        ],
        "title": "CreateDepositResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "WalletEntryPublic": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "amount": {
            "type": "integer",
            "title": "Amount"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "external_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Id"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "type",
          "amount",
          "currency",
          "created_at"
        ],
        "title": "WalletEntryPublic"
      },
      "WalletSnapshot": {
        "properties": {
          "balance": {
            "type": "integer",
            "title": "Balance"
          },
          "entries": {
            "items": {
              "$ref": "#/components/schemas/WalletEntryPublic"
            },
            "type": "array",
            "title": "Entries"
          }
        },
        "type": "object",
        "required": [
          "balance",
          "entries"
        ],
        "title": "WalletSnapshot"
      },
      "WithdrawRequest": {
        "properties": {
          "tokens": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Tokens",
            "description": "Number of tokens to withdraw"
          }
        },
        "type": "object",
        "required": [
          "tokens"
        ],
        "title": "WithdrawRequest"
      },
      "WithdrawResponse": {
        "properties": {
          "requested": {
            "type": "integer",
            "title": "Requested"
          },
          "refunded": {
            "type": "integer",
            "title": "Refunded"
          },
          "stripe_refunds": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Stripe Refunds"
          }
        },
        "type": "object",
        "required": [
          "requested",
          "refunded",
          "stripe_refunds"
        ],
        "title": "WithdrawResponse"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}